Used PCSX2's implementation for unaligned memory access
That, along with an adjustment to IOP DMA, fixed bugs in Disgaea, allowing it to read an IRX patch file successfully.
Also added PXOR

+++ src/core/iop/iop_dma.hpp
@@ -14,7 +14,7 @@ struct IOP_DMA_Chan_Control
 struct IOP_DMA_Channel
 {
     uint32_t addr;
-    uint16_t word_count;
+    uint32_t word_count;
     uint16_t block_size;
     IOP_DMA_Chan_Control control;
     uint32_t tag_addr;
